Adding the scrolling feature
I am implementing a feature that allows the user to click on the links in the navigation bar. Once a link is clicked the page scrolls, at a reasonable speed, to the particular section of the UI that is connected to that link. 

There are 3 parts of the header.js file that I want to learn more about. 1) the constructor function, 2) the componentDidMount code section and 3) the componentDidUnmount code piece. Specifically, I want to know more about exactly what they do.

Furthermore, there are a few things that I will change:
    1) This is not a huge project so I do not need several component files. I just need one index page containing all the code.
    2) I will remove some of the imports from the react-scroll package because I do not need them. I just need to figure out what 
        is truly relevant/irrelevant for this project.